build: extract 'TestUtils' to 'common' module
A number of circular dependencies on 'compat' meant
that all functionality could not be extracted from 'AdaptionUtil'

* We need to extract 'isRobolectric'
* isRunningAsUnitTest is an easy extraction
  * added in 7ada532c5cfb72626ec2c5ef348b7570de62453a
