+++ AnkiDroid/src/main/java/com/ichi2/anki/CollectionManager.kt
@@ -21,13 +21,13 @@ import android.content.Context
 import androidx.annotation.VisibleForTesting
 import androidx.annotation.WorkerThread
 import anki.backend.backendError
+import com.ichi2.anki.common.utils.android.isRobolectric
 import com.ichi2.anki.servicelayer.ValidatedMigrationSourceAndDestination
 import com.ichi2.anki.servicelayer.scopedstorage.MigrateEssentialFiles
 import com.ichi2.libanki.Collection
 import com.ichi2.libanki.Storage.collection
 import com.ichi2.libanki.importCollectionPackage
 import com.ichi2.utils.Threads
-import com.ichi2.utils.isRobolectric
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
@@ -40,7 +40,6 @@ import okio.withLock
 import timber.log.Timber
 import java.io.File
 import java.util.concurrent.locks.ReentrantLock
-import kotlin.concurrent.withLock
 
 object CollectionManager {
     /**
