+++ AnkiDroid/src/main/java/com/ichi2/utils/Permissions.kt
@@ -26,10 +26,10 @@ import androidx.annotation.RequiresApi
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
+import com.ichi2.anki.common.utils.android.isRobolectric
 import com.ichi2.compat.CompatHelper.Companion.getPackageInfoCompat
 import com.ichi2.compat.PackageInfoFlagsCompat
 import timber.log.Timber
-import java.lang.Exception
 
 object Permissions {
     const val MANAGE_EXTERNAL_STORAGE = "android.permission.MANAGE_EXTERNAL_STORAGE"
