+++ AnkiDroid/src/test/java/com/ichi2/anki/CardBrowserTest.kt
@@ -34,8 +34,10 @@ import com.ichi2.anki.browser.CardBrowserColumn
 import com.ichi2.anki.browser.CardBrowserViewModel
 import com.ichi2.anki.browser.CardBrowserViewModel.Companion.DISPLAY_COLUMN_1_KEY
 import com.ichi2.anki.browser.CardBrowserViewModel.Companion.DISPLAY_COLUMN_2_KEY
+import com.ichi2.anki.common.utils.isRunningAsUnitTest
 import com.ichi2.anki.dialogs.DeckSelectionDialog
-import com.ichi2.anki.model.CardsOrNotes.*
+import com.ichi2.anki.model.CardsOrNotes.CARDS
+import com.ichi2.anki.model.CardsOrNotes.NOTES
 import com.ichi2.anki.model.SortType
 import com.ichi2.anki.scheduling.ForgetCardsViewModel
 import com.ichi2.anki.servicelayer.NoteService
@@ -53,11 +55,12 @@ import com.ichi2.testutils.OS
 import com.ichi2.testutils.TestClass
 import com.ichi2.testutils.getSharedPrefs
 import com.ichi2.ui.FixedTextView
-import com.ichi2.utils.AdaptionUtil
 import com.ichi2.utils.LanguageUtil
 import io.mockk.every
 import io.mockk.mockkObject
+import io.mockk.mockkStatic
 import io.mockk.unmockkObject
+import io.mockk.unmockkStatic
 import kotlinx.coroutines.runBlocking
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
@@ -396,11 +399,11 @@ class CardBrowserTest : RobolectricTest() {
     @Test
     fun startupFromCardBrowserActionItemShouldEndActivityIfNoPermissions() {
         try {
-            mockkObject(AdaptionUtil)
+            mockkStatic(::isRunningAsUnitTest)
             mockkObject(IntentHandler)
 
             every { grantedStoragePermissions(any(), any()) } returns false
-            every { AdaptionUtil.isRunningAsUnitTest } returns false
+            every { isRunningAsUnitTest } returns false
 
             val browserController = Robolectric.buildActivity(CardBrowser::class.java).create()
             val cardBrowser = browserController.get()
@@ -408,7 +411,7 @@ class CardBrowserTest : RobolectricTest() {
 
             assertThat("Activity should be finishing", cardBrowser.isFinishing)
         } finally {
-            unmockkObject(AdaptionUtil)
+            unmockkStatic(::isRunningAsUnitTest)
             unmockkObject(IntentHandler)
         }
     }
