+++ common/src/main/java/com/ichi2/anki/common/utils/TestUtils.kt
@@ -1,6 +1,4 @@
 /*
- *  Copyright (c) 2024 David Allison <davidallisongithub@gmail.com>
- *
  *  This program is free software; you can redistribute it and/or modify it under
  *  the terms of the GNU General Public License as published by the Free Software
  *  Foundation; either version 3 of the License, or (at your option) any later
@@ -13,28 +11,23 @@
  *  You should have received a copy of the GNU General Public License along with
  *  this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+package com.ichi2.anki.common.utils
 
-package com.ichi2.anki.logging
-
-import com.ichi2.anki.BuildConfig
-import com.ichi2.anki.common.utils.android.isRobolectric
 import timber.log.Timber
 
-enum class LogType {
-    /** @see Timber.DebugTree */
-    DEBUG,
-
-    /** @see RobolectricDebugTree */
-    ROBOLECTRIC,
-
-    /** @see ProductionCrashReportingTree */
-    PRODUCTION;
-
-    companion object {
-        val value: LogType
-            get() {
-                if (!BuildConfig.DEBUG) return PRODUCTION
-                return if (isRobolectric) ROBOLECTRIC else DEBUG
-            }
+/** make default HTML / JS debugging true for debug build and disable for unit/android tests
+ * isRunningAsUnitTest checks if we are in debug or testing environment by checking if org.junit.Test class
+ * is imported.
+ * https://stackoverflow.com/questions/28550370/how-to-detect-whether-android-app-is-running-ui-test-with-espresso
+ */
+val isRunningAsUnitTest: Boolean
+    get() {
+        try {
+            Class.forName("org.junit.Test")
+        } catch (ignored: ClassNotFoundException) {
+            Timber.d("isRunningAsUnitTest: %b", false)
+            return false
+        }
+        Timber.d("isRunningAsUnitTest: %b", true)
+        return true
     }
-}
